Fix inconsitent `Menu` API for Linux

+++ src/platform_impl/linux/menu.rs
@@ -22,7 +22,7 @@ impl Menu {
 
     pub fn add_item<S: Into<String>, H: Into<Option<Hotkey>>>(
         &mut self,
-        _id: usize,
+        _id: u16,
         _name: S,
         _key: H,
     ) {
